+++ src/main/java/com/github/steveice10/mc/protocol/data/game/world/particle/DustParticleData.java
@@ -6,8 +6,8 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class DustParticleData implements ParticleData {
-    private final double red; // 0 - 1
-    private final double green; // 0 - 1
-    private final double blue; // 0 - 1
+    private final float red; // 0 - 1
+    private final float green; // 0 - 1
+    private final float blue; // 0 - 1
     private final float scale; // clamped between 0.01 and 4
 }
