+++ src/main/java/com/github/steveice10/mc/protocol/data/game/world/particle/ParticleData.java
@@ -12,19 +12,19 @@ public interface ParticleData {
             case BLOCK:
                 return new BlockParticleData(in.readVarInt());
             case DUST:
-                double red = in.readDouble();
-                double green = in.readDouble();
-                double blue = in.readDouble();
+                float red = in.readFloat();
+                float green = in.readFloat();
+                float blue = in.readFloat();
                 float scale = in.readFloat();
                 return new DustParticleData(red, green, blue, scale);
             case DUST_COLOR_TRANSITION:
-                red = in.readDouble();
-                green = in.readDouble();
-                blue = in.readDouble();
+                red = in.readFloat();
+                green = in.readFloat();
+                blue = in.readFloat();
                 scale = in.readFloat();
-                double newRed = in.readDouble();
-                double newGreen = in.readDouble();
-                double newBlue = in.readDouble();
+                float newRed = in.readFloat();
+                float newGreen = in.readFloat();
+                float newBlue = in.readFloat();
                 return new DustColorTransitionParticleData(red, green, blue, scale, newRed, newGreen, newBlue);
             case FALLING_DUST:
                 return new FallingDustParticleData(in.readVarInt());
@@ -41,19 +41,19 @@ public interface ParticleData {
                 out.writeVarInt(((BlockParticleData) data).getBlockState());
                 break;
             case DUST:
-                out.writeDouble(((DustParticleData) data).getRed());
-                out.writeDouble(((DustParticleData) data).getGreen());
-                out.writeDouble(((DustParticleData) data).getBlue());
+                out.writeFloat(((DustParticleData) data).getRed());
+                out.writeFloat(((DustParticleData) data).getGreen());
+                out.writeFloat(((DustParticleData) data).getBlue());
                 out.writeFloat(((DustParticleData) data).getScale());
                 break;
             case DUST_COLOR_TRANSITION:
-                out.writeDouble(((DustParticleData) data).getRed());
-                out.writeDouble(((DustParticleData) data).getGreen());
-                out.writeDouble(((DustParticleData) data).getBlue());
+                out.writeFloat(((DustParticleData) data).getRed());
+                out.writeFloat(((DustParticleData) data).getGreen());
+                out.writeFloat(((DustParticleData) data).getBlue());
                 out.writeFloat(((DustParticleData) data).getScale());
-                out.writeDouble(((DustColorTransitionParticleData) data).getNewRed());
-                out.writeDouble(((DustColorTransitionParticleData) data).getNewGreen());
-                out.writeDouble(((DustColorTransitionParticleData) data).getNewBlue());
+                out.writeFloat(((DustColorTransitionParticleData) data).getNewRed());
+                out.writeFloat(((DustColorTransitionParticleData) data).getNewGreen());
+                out.writeFloat(((DustColorTransitionParticleData) data).getNewBlue());
                 break;
             case FALLING_DUST:
                 out.writeVarInt(((FallingDustParticleData) data).getBlockState());
